Remove tf.io.loadWeights()
BREAKING

The functionality has been superceded by IOHandlers.
